feat: warn if replugged install is invalid

+++ index.ts
@@ -1,6 +1,7 @@
 import RikkaPlugin from "@rikka/Common/entities/Plugin";
 import { registerURLCallback } from "@rikka/modules/browserWindowtils";
-import { copyFileSync } from "fs";
+import { ipcRenderer } from "electron";
+import { copyFileSync, existsSync } from "fs";
 import { join } from "path";
 import { rpPath } from "./constants/pc";
 import manifest from "./manifest.json";
@@ -22,6 +23,16 @@ export default class RepluggedCompat extends RikkaPlugin {
   }
 
   inject() {
+    if (!existsSync(this.rpInstallPath)) {
+      ipcRenderer.invoke("RPCOMPAT_SHOW_DIALOG", {
+        title: "Replugged Compat",
+        type: "error",
+        message: "Unable to find a valid Replugged installation.",
+        detail: "Please download and install Replugged following the instructions in the README.",
+        buttons: ["OK"],
+      });
+      return;
+    }
     copyFileSync(join(__dirname, "preload.js"), join(this.rpInstallPath, "src", "preload.js"));
 
     this.log("Loading Replugged");
